#1904 Added option to filter the list on dungeon ID & validation for it
