+++ app/Http/Requests/Api/V1/APIFormRequest.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Http\Requests\Api\V1;
+
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Http\Exceptions\HttpResponseException;
+
+abstract class APIFormRequest extends FormRequest
+{
+    public function failedValidation(Validator $validator)
+    {
+        throw new HttpResponseException(response()->json([
+            'success' => false,
+            'message' => 'Validation errors',
+            'data'    => $validator->errors(),
+        ]));
+    }
+}
